feat(app): Add gripper error route to Error Recovery (#16396)
Partially closes EXEC-723

This commit adds most of the new gripper error flow to Error Recovery. See the linked ticket for what's still missing and what's blocking completion.

This new flow deviates from other recovery flows in the following ways:

* We have to special case the future gripper jaw release command not to route us to the door open CTAs.
* We have a command that operates on a countdown. In practice, this is tricky to account for given the current ER FE architecture. The current solution does a good enough job of handling this special case (we now route RecoveryInProgress, and this view has logic for actually executing the gripper command). There's no great silver bullet to make this work, but this solution is probably the most straightforward method for handling this. I think if we do any more of "countdown until execution" command type logic, then we can spend the time to handle these scenarios more elegantly.
* We have new gripper errors views!
* As ER has settled from 8.0, it's easier to see which views can be and should be refactored. We have a lot of "retry step" and "skip to next step" views that could be consolidated and used with the new gripper flows, so that's done here.

Note that this recovery route is not publicly accessible yet. For testing purposes, it's accessible with a bit of hacking, but it will be more thoroughly tested with the commit that actually closes the ticket.