feat(shared-data): add command schema 10 (#16401)
This is a new default command schema that we'll use for implementing
static meniscus-relative pipetting.

There are a couple odd changes in step-generation to look at because the
command types finally changed in a way that was incompatible with the
schema v6 types there. However, there's not an incompatible change in
the commands, it's just that the types were inappropriately wide.