feat(api,shared-data): add runExtendedProfile
This is a thermocycler command that accepts a structured sequence of
temperatures and wait times instead of a flat list.

A profile is a list of either steps or cycles; cycles are a repetition
count plus a list of steps.

The actual _functionality_ of the command is exactly the same as
thermocycler/runProfile; the only change is pushing down the structure
of the information, which is currently lost when creating a runProfile
command and thus unavailable to protocol clients for display.