fix(app): update thermocycler command text
This PR does a couple things with thermocycler command text.

First, it changes the phrasing of the thermocycler/runProfile command
text to make it more clear that no cycles are involved: it's just a big
list of steps.

Second, it adds a command text for the new
thermocycler/runExtendedProfile command, which does have the structured
data necessary to render things as cycles. This unfortunately makes it a
multilevel list, thus worsening the annoying problems with it.

Closes RQA-2771