feat(robot-server): Add `GET` `runs/:runId/currentState` (#16402)
Works toward RSQ-161

To support partial tip during error recovery during certain flows, the server must expose details about the current pipetting tip layout for the current run. NozzleMap contains details the client is interested in, but we really shouldn't return the entirety of NozzleMap, since it is A) quite large and B) contains a lot of info irrelevant from a client perspective. To that end, we return an ActiveNozzleLayout derived from the NozzleMap for the pipette of interest.

This commit returns nozzle data on a new endpoint, /runs/:runId/currentState.