feat(robot-server): Wire up global error recovery setting (#16416)
## Overview

Closes EXEC-719.

## Changelog

* Connect the new `PATCH /errorRecovery/settings` HTTP API so it
actually affects run behavior. This closes EXEC-719.
    * One quirk to my implementation:
    
This will not affect preexisting runs, *unless* you *also* do `PUT
/runs/{id}/errorRecoveryPolicy`. In effect, `PUT
/runs/{id}/errorRecoveryPolicy` refreshes the run's snapshot of
`/errorRecovery/settings`. This is probably confusing; it would be
better if `PATCH /errorRecovery/settings` immediately affected the
preexisting run, or if `PUT /runs/{id}/errorRecoveryPolicy` did not have
that refresh effect.
    
I don't think this matters in practice because this button will be
disabled or inaccessible in our UIs while there's an active run.

Also fix a couple of nearby, unticketed bugs:

* If a client gave us an empty error recovery policy, we were no-opping.
Instead, we want to overwrite the current policy with the new empty one.
* The logic in `error_recovery_mapping.py` was unintentionally enabling
error recovery on OT-2 robots if the policy was specific enough. We want
it to always be disabled on OT-2s.