fix(components): fix logic for whether TC is selected in DeckConfigurator
In our `DeckConfigurator` component, there is an insufficient equality check for selected cutout ID
with configured thermocycler cutout. The situation arises when a thermocycler is configured in
cutout A1 and the selected cutout ID is B1, or vice versa. These point to the thermocycler being
configured in the same physical location, so we should check whether both selected and configured
cutout IDs are included in thermocycler cutouts ([cutoutA1, cutoutB1]).

Closes RQA-3237