fix(components): fix logic for whether TC is selected in DeckConfigurator (#16476)
In our `DeckConfigurator` component, there is a leaky equality check for
selected cutout ID with configured thermocycler cutout. The situation
arises when a thermocycler is configured in cutout A1 and the selected
cutout ID is B1, or vice versa. These point to the thermocycler being
configured in the same physical location, so we should check whether
both selected and configured cutout IDs are included in thermocycler
cutouts ([cutoutA1, cutoutB1]).