refactor: actually save stuff to disk
Awkwardly enough, we've been saving all our "persistently cached update
data" to a ramdisk. This is bad! It's bad because (a) it gets cleared on
reboots, so it's not actually persistently stored and (b) it's a couple
gigabytes, which for sure shouldn't be on a ramdisk.

This happens because the update storage was in the electron userdata
path, which for some reason (probably we have the wrong setting of some
XDG variable) was resolving to live in /tmp.

Instead, explicitly set it (if in production) to the same directory in
/data that we store config data, since that's also conveniently what
gets cleared when you factory reset the robot.

Since we're now using the app path for that, we can alter our logging
configuration and our electron-store config configuration to use those
paths instead of hardcoding. Sweet!