fix(shared-data): fix broken RTP choice range formatter util
We developed a utility function `orderRuntimeParameterRangeOptions` to order choice-type parameters
in the `ParametersTable` and `ProtocolDetails` components. However, the util incorrectly assumed
that the RTP choice array passed to the utility was of length 2 when it could also be of length 1.
This PR filters the array of any length rather than explicitly indexing its second element to fix
the bug.

Closes AUTH-929