feat(app): add Tip Drop Failed flow to Error Recovery (#16496)
Closes EXEC-722

This PR wires up the FE implementation of the "Tip Drop Failed" Error Recovery flow.

This flow is straightforward. The only real refactor is that the copy for recovery options is now special-cased, so instead of "Retry step", it's, for example, "Retry dropping tip".

Note that the feature is not fully implemented until EXEC-764 closes.