Implement drop_tip()'s state updates in terms of other public methods, where possible.
This is not intended to have any behavioral change.

The goal of this is to make it easier for callers (i.e. Protocol Engine) to see what they have to do in order to replicate the total effect of drop_tip(). And to make it easier for human code reviewers to confirm that that's done correctly.