fix: next-tickify dispatches
Now that we were looping back through action handlers, there was a risk
of an infinite recursion if an action handler dispatched an action that
it handled by dispatching an action. We hate those. This will make it
just an infinite sequence of events which at least won't crash the vm