fix(app): don't re-open drop tip wizard flows if unrendered (#16540)
Closes RQA-3287

In #16490, I made the disable/enable drop tip wizard states more explicit. However, I also removed the logic to toggle() the wizard if it abruptly closes, which sometimes happens if the pipette card unrenders. This just adds the unrender logic back and makes it more explicit: if the flow does close abruptly, it's important we do clean up the maintenance run, too.