feat(protocol-designer): update StepSummary render logic and copy (#16512)
This PR updates when StepSummaries show and updates copy to reflect
final decisions. The intended behavior is to only show StepSummary when
a step is hovered or selected, but not when the step form is open.
Hovered StepSummaries take precedence over those of selected steps. I
also add an optional second element to StepSummary that will display the
step's notes if they exists.

Closes AUTH-886