feat(hardware): refactor tool_sensors to simplify and remove csvs (#16462)
<!--
Thanks for taking the time to open a Pull Request (PR)! Please make sure
you've read the "Opening Pull Requests" section of our Contributing
Guide:


https://github.com/Opentrons/opentrons/blob/edge/CONTRIBUTING.md#opening-pull-requests

GitHub provides robust markdown to format your PR. Links, diagrams,
pictures, and videos along with text formatting make it possible to
create a rich and informative PR. For more information on GitHub
markdown, see:


https://docs.github.com/en/get-started/writing-on-github/getting-started-with-writing-and-formatting-on-github/basic-writing-and-formatting-syntax

To ensure your code is reviewed quickly and thoroughly, please fill out
the sections below to the best of your ability!
-->

# Overview
This PR removes much of the complexity in tool_sensors around the liquid
probe routine.
Instead of having many different options of setting the bindings and
building messages, this consolidates all of the previous options by only
commanding the firmware one way, forwarding the sensor information to a
new log file, and providing a way for the system to optionally grab the
raw sensor data if they wish which is required for the hardware-testing
scripts and possibly for future features.

We no longer write CSV files to the robot through this system and have
removed all of the coded in paths for them. It also removes the
"OutputOptions" data type that previously controlled the behavior
variants.

Before each variant had it's own quirks to behavior so now machine
actions will always be the same regardless of how the tool_sensors
method is called.

This also adds a new can message that sends sensor data in batches
instead of 1 by 1 so we don't choke up the can bus as much.

<!--
Describe your PR at a high level. State acceptance criteria and how this
PR fits into other work. Link issues, PRs, and other relevant resources.
-->

## Test Plan and Hands on Testing

<!--
Describe your testing of the PR. Emphasize testing not reflected in the
code. Attach protocols, logs, screenshots and any other assets that
support your testing.
-->
## Changelog

<!--
List changes introduced by this PR considering future developers and the
end user. Give careful thought and clear documentation to breaking
changes.
-->

## Review requests

<!--
- What do you need from reviewers to feel confident this PR is ready to
merge?
- Ask questions.
-->

## Risk assessment

<!--
- Indicate the level of attention this PR needs.
- Provide context to guide reviewers.
- Discuss trade-offs, coupling, and side effects.
- Look for the possibility, even if you think it's small, that your
change may affect some other part of the system.
- For instance, changing return tip behavior may also change the
behavior of labware calibration.
- How do your unit tests and on hands on testing mitigate this PR's
risks and the risk of future regressions?
- Especially in high risk PRs, explain how you know your testing is
enough.
-->