feat: Opentrons ai client landing page (#16552)
# Overview

This PR refactors Opentrons AI Client to match the new design. It also
adds the mixpanel files for tracking analytics.


![image](https://github.com/user-attachments/assets/cebc0744-c791-4426-9dda-f7b1121a4e19)

## Test Plan and Hands on Testing

Still wip, but tested manually the landing page and its buttons, and if
mixpanel is tracking the events.

## Changelog

 - New Opentron AI client flow and Landing page.
 - Add mixpanel analytics.

## Review requests

- Landing page.
- Verify if the approach for Mixpanel implementation is ok.

## Risk assessment

It breaks the current Opentrons AI client, as it's being remade from the
ground up.