fix(app): avoid reodering in robot dashboard
Here are some interesting facts:
- Array.sort() mutates
- Array.reverse() mutates
- react query caches data

These facts combined to mean that in the ODD robot dashboard, protocols
view, and run history, we would almost always rerender multiple times
with rapidly-reordering data and move stuff out from under someone
trying to interact with us. I'm actually pretty surprised that these
would usually end up in the right order instead of exactly reversed.

The fix for this is to quit using those methods, and the way to do that
is to mark the data read only. We should do this for all data returned
by a react-query hook IMO.

Closes RQA-3422