fix(app): remove some dangerous reverse()s
There weer a couple command texts that were doing a `.reverse()` of the
commands array. This is O(N) which is gross and also might be cached
which would cause bugs, so replace it with some uses of findLast() that
definitely weren't inspired by wanting to do some fun typing exercises.