feat(protocol-designer): introduce LiquidEntities and update ingredients JSON key to match (#17385)
closes AUTH-1382

This PR introduces `liquidEntities` to `InvariantContext` so
step-generation will have access to liquids and the liquid `pythonName`.
As a result, the previous selector that stored liquid info,
`allIngredientNamesIds` has been deprecated.

Additionally, `pythonName` is assigned during liquid creation via the
`EDIT_LIQUID_GROUP` action and if a user deletes a liquid, the
`deleteLiquidGroup` thunk is called which: deletes the liquid, edits all
subsequent liquids with the new `liquidEntities` object.