feat(protocol-designer): add designerApplication labware, pipettes, m… (#17432)
…odules keys

closes AUTH-1407

This PR introduces `designerApplication` top level keys for `modules`,
`labware`, and `pipettes`. Then when you import a protocol, the migration will grab that info from the load commands and the `load-file` reducer will use that info to populate the `labwareEntities`, `pipetteEntities`, and `moduleEntities`.

Additionally, this pr deprecates the `defaultValues` key that was not in use at all