feat(protocol-designer, step-generation, app): introduce pythonName to each entity (#17463)
closes AUTH-1383

For python/pd interop, we want to add a `pythonName` to `liquidEntity`,
`labwareEntity`, `moduleEntity`, and `pipetteEntity`. This PR does that
by adding them in redux, that way, the `pythonName` is not polluting the
JSON structure. For Quick Transfer, the `pythonName` is generated in
`generateQuickTransferArgs`.

The python name patterning is as follows:
- modules: moduleType minus `Type` in snake case plus the number so
`heater_shaker_module_1` and then `heater_shaker_module_2` for 2
heater-shaker modules
- pipettes: `pipette_left` or `pipette_right`
- labware: the labware displayCategory in snake case plus the number. So
2 well plates and 1 reservoir would be `well_plate_1`, `well_plate_2`,
`reservoir_1`
- liquids: just liquid with the number so `liquid_1`, `liquid_2`, etc.