refactor(step-generation): make dropTipInTrash a compound command instead of a util function (#17726)
# Overview

Previously, we had a util function in `movableTrashCommandsUtil.ts` to
drop a tip into the trash bin. This PR changes it to a compound command
to fit the pattern of our other `CommandCreators`, and to make it easier
to implement Python generation for drop-tip later. AUTH-1541

While working on this refactoring, we discovered that the previous
implementation was not dropping the tip at the end of `consolidate()`
like it should have because the previous implementation was passing a
stale copy of `prevRobotState` to the util function. After this
refactoring, we do drop tip at the end of `consolidate` as intended, so
this PR also updates `consolidate.test.ts`.

## Test Plan and Hands on Testing

I'm relying on the existing unit tests to check that `dropTipInTrash`
continues to work as expected.

I migrated the `dropTipInTrash` test from
`movableTrashCommandsUtil.test.ts` to `dropTipInTrash.test.ts` and
reformatted it to be a compound command test.

I updated `consolidate.test.ts` to expect drop-tip at the end of
`consolidate()`, which is the behavior we expect.

## Risk assessment

Medium low. In isolated tests, this change will make us drop-tip where
we were failing to do so before. But in practice, users will not see a
difference, because we've also always had a separate "eager" drop-tip
mechanism that added a drop-tip to protocols even when we were failing
to generate one before.