fix(shared-data): Zero out cornerOffsetFromSlot in opentrons_tough_pcr_auto_sealing_lid (#17723)
## Overview

Currently, `opentrons_tough_pcr_auto_sealing_lid` has a
`cornerOffsetFromSlot` of (0, 0, -0.71). So, hypothetically, if you
placed it in a deck slot, its bottom would be 0.71 mm "underground,"
below the deck. This seems wrong.

It seems like this was part of a snafu related to gripper offsets.
Investigation is ongoing in EXEC-1268, which you should read for some
background. But it seems clear at this point that
`cornerOffsetFromSlot`, at least, ought to be zero.

## Test Plan and Hands on Testing

* [x] Robot behavior is unchanged.
* As described in EXEC-1268, the robot always behaved as if this offset
was (0, 0, 0) because of an unrelated bug, EXEC-1285.

## Changelog

* Zero out `cornerOffsetFromSlot` in the as-yet-unreleased v2 of this
definition.
* Zero out `cornerOffsetFromSlot` in v1 of this definition, which *has*
already been released.

This is "bad"; we should usually not change definitions retroactively
like this. I think it's justified in this case because someday we will
fix EXEC-1285 (intentionally or not), so if we leave the offset in
place, we're setting ourselves up for a change in behavior in the
future. Better to nip it in the bud now, when this feature is still new.
  
Also, @CaseyBatten changed the definition in #17683, and if he can do it
then so can I. :P

But this isn't a hill I'll die on. I'm happy to keep the offset in the
v1 definition if we accept the risk of it making more of a mess in the
future.

## Risk assessment

Low.