feat(step-generation): generate Python for dropTipInTrash compound command (#17740)
# Overview

This is the first compound command we're generating Python for (and it's
the last piece we need to get a runnable Python protocol for simple
transfers AUTH-1093).

As discussed earlier, our strategy for compound commands that map to a
single Python command is that we'll add an extra `CommandCreator` that
emits the single Python command. The extra `CommandCreator` does not
emit any JSON commands, so the JSON behavior is unchanged.

## Test Plan and Hands on Testing

Added unit tests.

I also tried this in my private branch, and confirmed that the syntax is
correct with `simulate`.

## Review requests

This implementation is brittle because the PAPI only allows you to vary
the drop location within the trash bin (which we want) when you
**DON'T** specify the trash bin location in the call to
`pipette.drop_tip(location=...)`. So this would only work when there is
only 1 trash bin, which is true for now. (See AUTH-1372 for more
discussion.)

## Risk assessment

Low. Python export is behind a feature flag.