feat(step-generation): use dropTipInTrash compound command in generateRobotStateTimeline (#17746)
# Overview

`generateRobotStateTimeline()` is responsible for emitting the final
drop-tip at the end of step like Transfer or Mix. (We call this "eager
tip dropping.")

Previously, `generateRobotStateTimeline()` manually generated the
`moveToAddressableAreaForDropTip` + `dropTipInPlace` commands for
dropping the tip into the trash bin at the end of the step.

This PR changes it to use the `dropTipInPlace` compound command instead.
Since the compound command implements Python generation, after this
change, eager tip dropping will emit the Python `drop_tip()` at the end
of the step, which is what we want. AUTH-1093

## Test Plan and Hands on Testing

I'm relying on existing unit tests to confirm that the JSON commands
haven't changed.

I added a new Python test for `generateRobotStateTimeline()` and
confirmed that we now emit the eager-tip-dropping `drop_tip()` Python
command at the end of each step.

I also examined the "Export Python" file from PD running in this branch.

## Risk assessment

Low. This should not affect any existing behavior for JSON protocols,
and Python protocols are hidden behind a feature flag.