fix(shared-data): fix definition of PickUpTipParams, should not inherit from TouchTipParams (#17777)
# Overview

`PickUpTipParams` was defined to be the same as `TouchTipParams`, but
that doesn't seem to make sense. `TouchTipParams` has fields like
`speed`, which isn't applicable to the `pickUpTip` command.

@sfoster1 It looks like you originally authored this code in PR #13848.
Was this a typo?

@ncdiehl11 Hm, or was it your change that added `speed` to
`PickUpTipParams` in PR #17545?

In any case, I'm trying to untangle the definition of `PickUpTipParams`
from `TouchTipParams` as part of the PD Python export implementation.

## Test Plan and Hands on Testing

I'm relying on our CI tests to see what breaks.

## Risk assessment

Just a type definition change, should not affect user-visible behavior.