Annotate mode="after" validators.
These can assume that the value they are given is already of the correct type (as declared on the field), since they run after Pydantic's validation.

This catches one bug where we were not correctly handling the `null` in `POST /settings/pipettes/:id {"fields": null}`, which the request model theoretically allows. Fix it because it's easy.