refactor(robot-server): Type-check robot_server/service/* (#17810)
## Overview

Various type-checking fixes pulled out of EXEC-1342.

## Test Plan and Hands on Testing

* [x] `PATCH /settings/pipettes/:id` still works
* [x] `POST /robot/move` still works, particularly with `"target":
"mount"`, and raises an error if z<30.
* [x] `POST /robot/home` still works, particularly with `"target":
"pipette"`, and raises an error if you do not specify `"mount"` in that
case.
* [x] `POST /settings/log_level/local` still works.
* [x] `GET /wifi/keys` and `POST /wifi/keys` still work.

## Changelog

Mostly adding obvious return type annotations. But there are some
behavioral changes in here, around Pydantic validators and FastAPI
endpoint functions, because it seemed just as easy to fix them as to
annotate them with `#type: ignore`s.

See the individual commit messages in the PR for details.

## Risk assessment

Medium. We need to be careful about the Pydantic validators and FastAPI
endpoint functions.