feat(app): Use notifications for labware offsets instead of polling (#17832)
## Overview

The client-side half of EXEC-1342.

## Test Plan and Hands on Testing

* [x] I guess at this point we can turn on the feature flag and try this
out by entering LPC on one client and changing offsets on another?
  * Tested with a merge of this PR, #17853, #17817, and #17850.

## Changelog

The client was polling `POST /labwareOffsets/searches`. This switches it
to only happen when the client receives a notification from the server
that offsets have changed, with a polling fallback. I'm entirely
cargo-culting the patterns from the existing `useNotify___()` hooks.

## Risk assessment

Medium. I don't think we have a good way to integration-test this kind
of thing, so there are risks of things like missed updates, or
inadvertent polling.